backend/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
python-dotenv
requests